requirements.txt
rich
questionary
ffmpeg-python